NEW  - three new classes have been added for support of different reaction group simulations. They are currently unused, but serve as a base for future changes.   + reactiongroup.h/cpp: this is an abstract class from which the two other classes inherit. This class can by itself perform reactions and update reaction rates but is unable to schedule reaction timings. Its API gives access to next reaction time and enables to perform the next scheduled reaction, but scheduling is done by classes inheriting this abstract class.  + constantrategroup.h/cpp: this class simulates reactions under a constant rate hypothesis. The reaction performing part is handled by the base class ReactionGroup. ConstantRateGroup does the scheduling over a given time interval. This is efficient as under constant rate, it can be done once at the beginning of the time interval.  + updatedrategroup.h/cpp: this class simulates reactions that constantly update their rate. As its sister class, its job is to do the scheduling, ReactionGroup handling the reaction performing/rate updating. This is slightly more sophisticated as rescheduling needs to be done anytime something happens elsewhere.
DELETED

MAJOR
 - randomhandler.h/cpp: new functionality were added to support functionalities needed by ConstantRateGroup.
  + a draw_poisson member enables drawing a random number according to a poisson distribution.
  + a draw_multiple_indices member was added that extends the draw_index functionality. The idea was to draw an index on a biased wheel, but it is possible to optimize this function if several values need to be drawn on the SAME biased wheel. More specifically, finding the indices on the wheel can be performed quicker when looking for multiple values at once. This function is performed by the private member find_multiple_indices which relies on a sort function given by the private member sorted_indices (the latter necessating a helper class added within the RandomHandler private space).

MINOR
 - forwarddeclarations.h and Makefile.am were updated to include the new classes and slightly reorganized for readibility.

BUGFIXES
 - solver.h/cpp: _number_reactions_performed was declared as a double instead of an int

+++ src/solver.h
@@ -116,7 +116,7 @@ class Solver
    * @brief Accessor to the number of reactions that have occurred.
    * @return Number of reactions that have occurred from the beginning of simulation.
    */
-  double number_reactions_performed (void) const;  
+  int number_reactions_performed (void) const;  
   
 
   // ==========================
@@ -177,7 +177,7 @@ private:
   double _t;
 
   /** @brief Number of reactions that have been performed. */
-  double _number_reactions_performed;
+  int _number_reactions_performed;
 
   /** @brief Random handler used for determining next reaction. */
   RandomHandler _random_handler;
@@ -211,7 +211,7 @@ inline double Solver::time (void) const
   return _t;
 }
 
-inline double Solver::number_reactions_performed (void) const 
+inline int Solver::number_reactions_performed (void) const 
 {
   return _number_reactions_performed;
 }
